oi oi oi

+++ app/src/main/java/com/example/project/rakFragment.kt
@@ -12,6 +12,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.EditText
+import android.widget.Toast
 import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -98,8 +99,29 @@ class rakFragment : Fragment() {
                             commit()
                         }
                     }else if(which == 1){
-                        database.userDao().deleteSayur(listSayur[position])
                         getData(getIdUser?:0)
+                        AlertDialog.Builder(requireContext())
+                            .setTitle("Delete " + listSayur[position].nama)
+                            .setPositiveButton(
+                                "HAPUS",DialogInterface.OnClickListener{
+                                        dialog, which ->
+                                    database.userDao().deleteSayur(listSayur[position])
+                                    listSayur.removeAt(position)
+                                    adapterP.notifyDataSetChanged()
+                                    dialog.dismiss()
+                                    Toast.makeText(
+                                        requireContext(),
+                                        "Data Successfully Deleted",
+                                        Toast.LENGTH_SHORT
+                                    ).show()
+
+                                }
+                            )
+                            .setNegativeButton(
+                                "BATAL",DialogInterface.OnClickListener { dialog, which ->
+                                    dialog.dismiss()
+                                }
+                            ).show()
                     }else if(which == 2){
                         val bundle = Bundle()
                         bundle.putInt("sayurId",listSayur[position].uidSayur?:0)
